6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	char d;
 	float e;
 
-	print("Size of an int: %d byte(s)\n", sizeof(a);
-	print("Size of a long int: %d byte(s)\n", sizeof(b);
-	print("Size of a long long int: %d byte(s)\n", sizeof(c);
-	print("Size of a char: %d byte(s)\n", sizeof(d);
-	print("Size of a float: %d byte(s)\n", sizeof(e);
+	print("Size of an int: %d byte(s)\n", (unsigned long)sizeof(a));
+	print("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(b));
+	print("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(c));
+	print("Size of a char: %d byte(s)\n", (unsigned long)sizeof(d));
+	print("Size of a float: %d byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
